Moved Config.loadClassPath() to Core.loadClassPath() to eliminate a circular import
