+++ TileStache/Config.py
@@ -60,7 +60,7 @@ documentation for TileStache.Providers, TileStache.Core, and TileStache.Geograph
 
 import sys
 import logging
-from sys import stderr, modules
+from sys import modules
 from os.path import realpath, join as pathjoin
 try:
     from urllib.parse import urljoin, urlparse
@@ -323,7 +323,7 @@ def _parseConfigCache(cache_dict, dirpath):
             raise Exception('Unknown cache: %s' % cache_dict['name'])
 
     elif 'class' in cache_dict:
-        _class = loadClassPath(cache_dict['class'])
+        _class = Core.loadClassPath(cache_dict['class'])
         kwargs = cache_dict.get('kwargs', {})
         kwargs = dict( [(str(k), v) for (k, v) in kwargs.items()] )
 
@@ -457,7 +457,7 @@ def _parseConfigLayer(layer_dict, config, dirpath):
         provider_kwargs = _class.prepareKeywordArgs(provider_dict)
 
     elif 'class' in provider_dict:
-        _class = loadClassPath(provider_dict['class'])
+        _class = Core.loadClassPath(provider_dict['class'])
         provider_kwargs = provider_dict.get('kwargs', {})
         provider_kwargs = dict( [(str(k), v) for (k, v) in provider_kwargs.items()] )
 
@@ -475,45 +475,3 @@ def _parseConfigLayer(layer_dict, config, dirpath):
     layer.pixel_effect = pixel_effect
 
     return layer
-
-def loadClassPath(classpath):
-    """ Load external class based on a path.
-
-        Example classpath: "Module.Submodule:Classname".
-
-        Equivalent soon-to-be-deprecated classpath: "Module.Submodule.Classname".
-    """
-    if ':' in classpath:
-        #
-        # Just-added support for "foo:blah"-style classpaths.
-        #
-        modname, objname = classpath.split(':', 1)
-
-        try:
-            __import__(modname)
-            module = modules[modname]
-            _class = eval(objname, module.__dict__)
-
-            if _class is None:
-                raise Exception('eval(%(objname)s) in %(modname)s came up None' % locals())
-
-        except Exception as e:
-            raise Core.KnownUnknown('Tried to import %s, but: %s' % (classpath, e))
-
-    else:
-        #
-        # Support for "foo.blah"-style classpaths, TODO: deprecate this in v2.
-        #
-        classpath = classpath.split('.')
-
-        try:
-            module = __import__('.'.join(classpath[:-1]), fromlist=str(classpath[-1]))
-        except ImportError as e:
-            raise Core.KnownUnknown('Tried to import %s, but: %s' % ('.'.join(classpath), e))
-
-        try:
-            _class = getattr(module, classpath[-1])
-        except AttributeError as e:
-            raise Core.KnownUnknown('Tried to import %s, but: %s' % ('.'.join(classpath), e))
-
-    return _class
